chatbot-python/requirements.txt
cohere==5.8.1
django-cors-headers==4.4.0
django-filter==24.2
djoser==2.2.3
drf-yasg==1.21.7
fqdn==1.5.1
git-filter-repo==2.38.0
ijson==3.2.3
isoduration==20.11.0
jsonpointer==2.4
mysqlclient==2.2.4
nltk==3.8.1
notebook==7.1.3
pinecone-client==5.0.1
pip-chill==1.0.3
pydot==2.0.0
pyspellchecker==0.7.2
python-dotenv==1.0.1
scikit-learn==1.3.2
seaborn==0.13.2
uri-template==1.3.0
webcolors==1.13
